+++ src/main/java/io/conduit/SourceRecordConverter.java
@@ -1,0 +1,63 @@
+/*
+ * Copyright 2022 Meroxa, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.conduit;
+
+import com.google.protobuf.ByteString;
+import com.google.protobuf.Struct;
+import com.google.protobuf.util.JsonFormat;
+import io.conduit.grpc.Data;
+import lombok.SneakyThrows;
+import org.apache.kafka.connect.data.Schema;
+import org.apache.kafka.connect.source.SourceRecord;
+
+import static io.conduit.Utils.jsonConvSchemaless;
+
+abstract class SourceRecordConverter {
+    protected Data getKey(SourceRecord sourceRecord) {
+        if (sourceRecord.key() == null) {
+            return Data.newBuilder().build();
+        }
+        if (sourceRecord.keySchema() != null) {
+            return schemaValueToData(sourceRecord.keySchema(), sourceRecord.key());
+        }
+        throw new UnsupportedOperationException("keys without schemas not supported yet");
+    }
+
+    @SneakyThrows
+    protected Data schemaValueToData(Schema schema, Object value) {
+        byte[] bytes = jsonConvSchemaless.fromConnectData("", schema, value);
+        if (schema.type() != Schema.Type.STRUCT) {
+            return Data.newBuilder()
+                    .setRawData(ByteString.copyFrom(bytes))
+                    .build();
+        }
+        // todo try improving performance
+        // Internally, Kafka's JsonConverter transforms the input value into a JsonNode, then into a bytes array.
+        // Then, we create a string from the bytes array, and let the Protobuf Java library parse the struct.
+        // This lets us quickly get the correct struct, without having to handle all the possible types,
+        // but is probably not as efficient as it can be.
+        // See: https://github.com/ConduitIO/conduit-kafka-connect-wrapper/issues/60
+        var outStruct = Struct.newBuilder();
+        JsonFormat.parser().ignoringUnknownFields().merge(
+                new String(bytes),
+                outStruct
+        );
+        return Data.newBuilder()
+                .setStructuredData(outStruct)
+                .build();
+    }
+}
