+++ src/test/java/io/conduit/KafkaToOpenCDCTest.java
@@ -1,0 +1,138 @@
+package io.conduit;
+
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.google.protobuf.Value;
+import io.conduit.grpc.Record;
+import lombok.SneakyThrows;
+import org.apache.kafka.connect.data.Schema;
+import org.apache.kafka.connect.data.SchemaBuilder;
+import org.apache.kafka.connect.data.Struct;
+import org.apache.kafka.connect.source.SourceRecord;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class KafkaToOpenCDCTest {
+    private KafkaToOpenCDC underTest;
+    private Schema valueSchema;
+    private Schema keySchema;
+    private Struct testValue;
+    private JsonNode testRecord;
+
+    @BeforeEach
+    public void setUp() {
+        underTest = new KafkaToOpenCDC();
+
+        valueSchema = new SchemaBuilder(Schema.Type.STRUCT)
+                .name("customers")
+                .field("id", Schema.INT32_SCHEMA)
+                .field("name", Schema.STRING_SCHEMA)
+                .field("interests", SchemaBuilder.array(Schema.STRING_SCHEMA))
+                .field("trial", SchemaBuilder.BOOLEAN_SCHEMA)
+                .field("balance", Schema.FLOAT64_SCHEMA)
+                .build();
+        keySchema = new SchemaBuilder(Schema.Type.STRUCT)
+                .name("customer_id_schema")
+                .field("id", Schema.INT32_SCHEMA)
+                .build();
+
+        testValue = new Struct(valueSchema)
+                .put("id", 123)
+                .put("name", "foobar")
+                .put("trial", true)
+                .put("balance", 33.44)
+                .put("interests", List.of("aaa", "bbb"));
+        testRecord = Utils.mapper.createObjectNode()
+                .put("id", 123)
+                .put("name", "foobar")
+                .put("trial", true)
+                .put("balance", 33.44)
+                .set("interests", Utils.mapper.createArrayNode().add("aaa").add("bbb"));
+    }
+
+    @Test
+    public void testFromKafkaSource_Null() {
+        assertNull(underTest.apply(null));
+    }
+
+    @SneakyThrows
+    @Test
+    public void testFromKafkaSource_WithValueSchema_NoKeySchema() {
+        var sourceRecord = new SourceRecord(
+                Map.of("test-partition", "test_table"),
+                Map.of("test-offset", 123456L),
+                "test-topic",
+                2,
+                valueSchema,
+                testValue
+        );
+        Record.Builder conduitRec = underTest.apply(sourceRecord);
+        assertNotNull(conduitRec);
+
+        // verify payload
+        var payload = conduitRec.getPayload().getAfter().getStructuredData();
+        assertMatch(testValue, payload);
+        // assert timestamp is within last second
+        long createdAt = Long.parseLong(conduitRec.getMetadataOrThrow(OpenCdcMetadata.CREATED_AT));
+        assertTrue(
+                createdAt / 1_000_000 > System.currentTimeMillis() - 1000
+        );
+        // verify key
+        assertFalse(conduitRec.getKey().hasRawData());
+        assertFalse(conduitRec.getKey().hasStructuredData());
+    }
+
+    @SneakyThrows
+    @Test
+    public void testFromKafkaSource_WithValueSchema_WithKeySchema() {
+        var sourceRecord = new SourceRecord(
+                Map.of("test-partition", "test_table"),
+                Map.of("test-offset", 123456L),
+                "test-topic",
+                2,
+                keySchema,
+                new Struct(keySchema).put("id", 123),
+                valueSchema,
+                testValue
+        );
+        Record.Builder conduitRec = underTest.apply(sourceRecord);
+        assertNotNull(conduitRec);
+
+        // verify payload
+        var payload = conduitRec.getPayload().getAfter().getStructuredData();
+        assertMatch(testValue, payload);
+        // assert timestamp is within last second
+        long createdAt = Long.parseLong(conduitRec.getMetadataOrThrow(OpenCdcMetadata.CREATED_AT));
+        assertTrue(
+                createdAt / 1_000_000 > System.currentTimeMillis() - 1000
+        );
+        // verify key
+        assertFalse(conduitRec.getKey().hasRawData());
+        assertTrue(conduitRec.getKey().hasStructuredData());
+        com.google.protobuf.Struct key = conduitRec.getKey().getStructuredData();
+        assertEquals(123, key.getFieldsOrThrow("id").getNumberValue());
+    }
+
+    private void assertMatch(Struct expected, com.google.protobuf.Struct payload) {
+        assertEquals(expected.get("id"), (int) payload.getFieldsOrThrow("id").getNumberValue());
+        assertEquals(expected.get("name"), payload.getFieldsOrThrow("name").getStringValue());
+        assertEquals(expected.get("trial"), payload.getFieldsOrThrow("trial").getBoolValue());
+        assertEquals(expected.get("balance"), payload.getFieldsOrThrow("balance").getNumberValue());
+        List<String> interestsExpected = expected.getArray("interests");
+        List<String> interestsActual = payload.getFieldsOrThrow("interests")
+                .getListValue().getValuesList()
+                .stream()
+                .map(Value::getStringValue)
+                .collect(Collectors.toList());
+        assertEquals(interestsExpected, interestsActual);
+    }
+}
