+++ src/test/java/io/conduit/TransformationsTest.java
@@ -1,8 +1,10 @@
 package io.conduit;
 
+import java.util.List;
+import java.util.UUID;
+
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.protobuf.ByteString;
-import com.google.protobuf.Value;
 import com.google.protobuf.util.JsonFormat;
 import io.conduit.grpc.Change;
 import io.conduit.grpc.Data;
@@ -11,22 +13,16 @@ import lombok.SneakyThrows;
 import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.SchemaBuilder;
 import org.apache.kafka.connect.data.Struct;
-import org.apache.kafka.connect.source.SourceRecord;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.stream.Collectors;
-
-import static io.conduit.Transformations.fromKafkaSource;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class TransformationsTest {
     private Schema valueSchema;
-    private Schema keySchema;
-    private Struct testValue;
     private JsonNode testRecord;
 
     @BeforeEach
@@ -39,100 +35,13 @@ public class TransformationsTest {
                 .field("trial", SchemaBuilder.BOOLEAN_SCHEMA)
                 .field("balance", Schema.FLOAT64_SCHEMA)
                 .build();
-        keySchema = new SchemaBuilder(Schema.Type.STRUCT)
-                .name("customer_id_schema")
-                .field("id", Schema.INT32_SCHEMA)
-                .build();
 
-        testValue = new Struct(valueSchema)
-                .put("id", 123)
-                .put("name", "foobar")
-                .put("trial", true)
-                .put("balance", 33.44)
-                .put("interests", List.of("aaa", "bbb"));
         testRecord = Utils.mapper.createObjectNode()
                 .put("id", 123)
                 .put("name", "foobar")
                 .put("trial", true)
                 .put("balance", 33.44)
                 .set("interests", Utils.mapper.createArrayNode().add("aaa").add("bbb"));
-    }
-
-    @Test
-    public void testFromKafkaSource_Null() {
-        assertNull(fromKafkaSource(null));
-    }
-
-    @SneakyThrows
-    @Test
-    public void testFromKafkaSource_WithValueSchema_NoKeySchema() {
-        var sourceRecord = new SourceRecord(
-                Map.of("test-partition", "test_table"),
-                Map.of("test-offset", 123456L),
-                "test-topic",
-                2,
-                valueSchema,
-                testValue
-        );
-        Record.Builder conduitRec = Transformations.fromKafkaSource(sourceRecord);
-        assertNotNull(conduitRec);
-
-        // verify payload
-        var payload = conduitRec.getPayload().getAfter().getStructuredData();
-        assertMatch(testValue, payload);
-        // assert timestamp is within last second
-        long createdAt = Long.parseLong(conduitRec.getMetadataOrThrow(OpenCdcMetadata.CREATED_AT));
-        assertTrue(
-                createdAt / 1_000_000 > System.currentTimeMillis() - 1000
-        );
-        // verify key
-        assertFalse(conduitRec.getKey().hasRawData());
-        assertFalse(conduitRec.getKey().hasStructuredData());
-    }
-
-    @SneakyThrows
-    @Test
-    public void testFromKafkaSource_WithValueSchema_WithKeySchema() {
-        var sourceRecord = new SourceRecord(
-                Map.of("test-partition", "test_table"),
-                Map.of("test-offset", 123456L),
-                "test-topic",
-                2,
-                keySchema,
-                new Struct(keySchema).put("id", 123),
-                valueSchema,
-                testValue
-        );
-        Record.Builder conduitRec = Transformations.fromKafkaSource(sourceRecord);
-        assertNotNull(conduitRec);
-
-        // verify payload
-        var payload = conduitRec.getPayload().getAfter().getStructuredData();
-        assertMatch(testValue, payload);
-        // assert timestamp is within last second
-        long createdAt = Long.parseLong(conduitRec.getMetadataOrThrow(OpenCdcMetadata.CREATED_AT));
-        assertTrue(
-                createdAt / 1_000_000 > System.currentTimeMillis() - 1000
-        );
-        // verify key
-        assertFalse(conduitRec.getKey().hasRawData());
-        assertTrue(conduitRec.getKey().hasStructuredData());
-        com.google.protobuf.Struct key = conduitRec.getKey().getStructuredData();
-        assertEquals(123, key.getFieldsOrThrow("id").getNumberValue());
-    }
-
-    private void assertMatch(Struct expected, com.google.protobuf.Struct payload) {
-        assertEquals(expected.get("id"), (int) payload.getFieldsOrThrow("id").getNumberValue());
-        assertEquals(expected.get("name"), payload.getFieldsOrThrow("name").getStringValue());
-        assertEquals(expected.get("trial"), payload.getFieldsOrThrow("trial").getBoolValue());
-        assertEquals(expected.get("balance"), payload.getFieldsOrThrow("balance").getNumberValue());
-        List<String> interestsExpected = expected.getArray("interests");
-        List<String> interestsActual = payload.getFieldsOrThrow("interests")
-                .getListValue().getValuesList()
-                .stream()
-                .map(Value::getStringValue)
-                .collect(Collectors.toList());
-        assertEquals(interestsExpected, interestsActual);
     }
 
     @Test
